Avoid UsageContext
UsageContext is a legacy type. We should be using SoftwareComponentVariant instead. We are still stuck
using UsageContext in some places since it is used in SoftwareComponentInternal#getUsages, however this
PR removes all other occurances of UsageContext where it is not strictly required to use. This PR will
make a future transition to a new public API for defining variants easier.

In general, this PR:
* Updates all terminology from using 'usage' to 'variant'. This includes renaming any implementations of UsageContext
to have Variant in their name. Variable and method names are updated where possible
* Remove constructor of DefaultSoftwareComponentVariant which accepts a configuration. Migrate usages to ConfigurationSoftwareComponentVariant.
* Add some documentation explaining difficulties of removing UsageContext altogether.